Organized DTO classes into new folders and name spaces.
