+++ FribergFastigheter.Client/AutoMapper/DtoToViewModelAutoMapperProfile.cs
@@ -1,16 +1,18 @@
 ﻿using AutoMapper;
 using FribergFastigheter.Client.Models;
-using FribergFastigheter.Shared.Dto;
-using FribergFastigheter.Shared.Dto.Statistics;
+using FribergFastigheter.Shared.Dto.Broker;
+using FribergFastigheter.Shared.Dto.BrokerFirm;
+using FribergFastigheter.Shared.Dto.Housing;
+using FribergFastigheter.Shared.Dto.Image;
 
 namespace FribergFastigheter.Client.AutoMapper
 {
-	/// <summary>
-	/// An auto mapper profile that contains mappings for converting DTO classes to view model classes.
-	/// </summary>
-	/// <!-- Author: Jimmie -->
-	/// <!-- Co Authors: -->
-	public class DtoToViewModelAutoMapperProfile : Profile
+    /// <summary>
+    /// An auto mapper profile that contains mappings for converting DTO classes to view model classes.
+    /// </summary>
+    /// <!-- Author: Jimmie -->
+    /// <!-- Co Authors: -->
+    public class DtoToViewModelAutoMapperProfile : Profile
 	{
 		/// <summary>
 		/// Constructor.
@@ -22,8 +24,6 @@ namespace FribergFastigheter.Client.AutoMapper
 			CreateMap<MunicipalityDto, MunicipalityViewModel>();
 			CreateMap<ImageDto, ImageViewModel>();
 			CreateMap<PaginationDto, PaginationViewModel>();
-			CreateMap<BrokerFirmStatisticsDto, BrokerFirmStatisticsViewModel>();
-			CreateMap<StatisticItemDto, StatisticItemViewModel>();
 
 			CreateMap<BrokerFirmDto, BrokerFirmViewModel>()
 				.ForMember(dest => dest.Logotype, opt => opt.PreCondition(x => x.Logotype != null));
