+++ FribergFastigheter.Client/AutoMapper/ViewModelToDtoAutoMapperProfile.cs
@@ -1,15 +1,19 @@
 ﻿using AutoMapper;
 using FribergFastigheter.Client.Models;
-using FribergFastigheter.Shared.Dto;
+using FribergFastigheter.Shared.Dto.Broker;
+using FribergFastigheter.Shared.Dto.BrokerFirm;
+using FribergFastigheter.Shared.Dto.Housing;
+using FribergFastigheter.Shared.Dto.Image;
+using FribergFastigheter.Shared.Dto.Login;
 
 namespace FribergFastigheter.Client.AutoMapper
 {
-	/// <summary>
-	/// An auto mapper profile that contains mappings for converting view model classes to DTO classes.
-	/// </summary>
-	/// <!-- Author: Jimmie -->
-	/// <!-- Co Authors: -->
-	public class ViewModelToDtoAutoMapperProfile : Profile
+    /// <summary>
+    /// An auto mapper profile that contains mappings for converting view model classes to DTO classes.
+    /// </summary>
+    /// <!-- Author: Jimmie -->
+    /// <!-- Co Authors: -->
+    public class ViewModelToDtoAutoMapperProfile : Profile
 	{
         /// <summary>
         /// Constructor.
