+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Broker/BrokerDetailsPage.razor.cs
@@ -1,7 +1,8 @@
 ﻿using AutoMapper;
 using FribergFastigheter.Client.Models;
 using FribergFastigheter.Client.Services.FribergFastigheterApi;
-using FribergFastigheter.Shared.Dto;
+using FribergFastigheter.Shared.Dto.Broker;
+using FribergFastigheter.Shared.Dto.Housing;
 using Microsoft.AspNetCore.Components;
 
 namespace FribergFastigheter.Client.Pages.BrokerFirmMember.Broker
