+++ FribergFastigheter.Client/Pages/BrokerFirmMember/Firm/BrokerFirmDetailsPage.razor.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using FribergFastigheter.Client.Models;
 using FribergFastigheter.Client.Services.FribergFastigheterApi;
-using FribergFastigheter.Shared.Dto;
+using FribergFastigheter.Shared.Dto.BrokerFirm;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Web;
