+++ FribergFastigheter.Client/Services/FribergFastigheterApi/HousingApiService.cs
@@ -1,4 +1,6 @@
-﻿using FribergFastigheter.Shared.Dto;
+﻿using FribergFastigheter.Shared.Dto.Broker;
+using FribergFastigheter.Shared.Dto.BrokerFirm;
+using FribergFastigheter.Shared.Dto.Housing;
 using System.Diagnostics;
 using System.Net;
 using System.Net.Http;
