+++ FribergFastigheter.Server/Controllers/HousingApi/HousingBrokerFirmController.cs
@@ -1,16 +1,17 @@
 ﻿using AutoMapper;
-using FribergFastigheter.Shared.Dto;
 using FribergFastigheter.Server.Data.Interfaces;
 using FribergFastigheter.Server.Services;
 using FribergFastigheterApi.Data.DatabaseContexts;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
+using FribergFastigheter.Shared.Dto.BrokerFirm;
+using FribergFastigheter.Shared.Dto.Error;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace FribergFastigheter.Server.Controllers.HousingApi
 {
-	[Route("housing-api")]
+    [Route("housing-api")]
 	[ApiController]
 	public class HousingBrokerFirmController : ControllerBase
 	{
