+++ FribergFastigheter.Shared/Dto/Broker/CreatedBrokerDto.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FribergFastigheter.Shared.Dto
+namespace FribergFastigheter.Shared.Dto.Broker
 {
     /// <summary>
     /// A DTO class that contains data for a created broker.
