+++ FribergFastigheter.Shared/Dto/Broker/RegisterBrokerDto.cs
@@ -1,24 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
-namespace FribergFastigheter.Shared.Dto
+namespace FribergFastigheter.Shared.Dto.Broker
 {
     /// <summary>
-    /// A DTO class that contains data for editing a broker.
+    /// A DTO class that holds data for creating a broker. 
     /// </summary>
     /// <!-- Author: Marcus -->
     /// <!-- Co Authors: -->
-    public class EditBrokerDto
+    public class RegisterBrokerDto
     {
         #region Properties
-
-        /// <summary>
-        /// The ID of the broker.
-        /// </summary>
-        public int BrokerId { get; set; }
 
         /// <summary>
         /// The description of the broker.
@@ -26,26 +17,35 @@ namespace FribergFastigheter.Shared.Dto
         public string Description { get; set; } = "";
 
         /// <summary>
-        /// The email of the broker.
+        /// The username/email of the broker.
         /// </summary>
+        [Required]
         public string Email { get; set; } = "";
 
         /// <summary>
         /// The first name of the broker.
         /// </summary>
+        [Required]
         public string FirstName { get; set; } = "";
 
         /// <summary>
         /// The last name of the broker.
         /// </summary>
+        [Required]
         public string LastName { get; set; } = "";
+
+        /// <summary>
+        ///The password for the broker.
+        /// </summary>
+        [Required]
+        public string Password { get; set; } = "";
 
         /// <summary>
         /// The phone number of the broker.
         /// </summary>
+        [Required]
         public string PhoneNumber { get; set; } = "";
 
         #endregion
     }
 }
-
