+++ FribergFastigheter.Shared/Dto/Broker/UpdateBrokerDto.cs
@@ -1,49 +1,43 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace FribergFastigheter.Shared.Dto
+namespace FribergFastigheter.Shared.Dto.Broker
 {
     /// <summary>
-    /// A DTO class that holds data for creating a broker. 
+    /// A DTO class that represents a broker.
     /// </summary>
     /// <!-- Author: Marcus -->
     /// <!-- Co Authors: -->
-    public class RegisterBrokerDto
+    public class UpdateBrokerDto
     {
         #region Properties
 
         /// <summary>
-        /// The description of the broker.
+        /// The firm that the broker belongs to.
         /// </summary>
-        public string Description { get; set; } = "";
+        public int BrokerFirmId { get; set; }
 
         /// <summary>
-        /// The username/email of the broker.
+        /// The ID of the broker.
         /// </summary>
-        [Required]
+        public int BrokerId { get; set; }
+
+        /// <summary>
+        /// The email of the broker.
+        /// </summary>
         public string Email { get; set; } = "";
 
         /// <summary>
         /// The first name of the broker.
         /// </summary>
-        [Required]
         public string FirstName { get; set; } = "";
 
         /// <summary>
         /// The last name of the broker.
         /// </summary>
-        [Required]
         public string LastName { get; set; } = "";
-
-        /// <summary>
-        ///The password for the broker.
-        /// </summary>
-        [Required]
-        public string Password { get; set; } = "";
-
         /// <summary>
         /// The phone number of the broker.
         /// </summary>
-        [Required]
         public string PhoneNumber { get; set; } = "";
 
         #endregion
