+++ FribergFastigheter.Shared/Dto/Login/LoginResponseDto.cs
@@ -5,14 +5,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace FribergFastigheter.Shared.Dto
+namespace FribergFastigheter.Shared.Dto.Login
 {
     /// <summary>
-    /// A DTO class for logging in a user.
+    /// A DTO class containing data for a login response.
     /// </summary>
     /// <!-- Author: Marcus, Jimmie -->
     /// <!-- Co Authors: -->
-    public class LoginDto
+    public class LoginResponseDto
     {
         #region Properties
 
@@ -23,10 +23,10 @@ namespace FribergFastigheter.Shared.Dto
         public string UserName { get; set; } = "";
 
         /// <summary>
-        /// The password.
+        /// The token.
         /// </summary>
         [Required]
-        public string Password { get; set; } = "";
+        public string Token { get; set; } = "";
 
         #endregion
     }
